remove code about MAPZEN because of service termination
