+++ Mobile/Maps/Maps.Tizen.Mobile/shared/res/Maps.png
