+++ Mobile/Maps/Maps.sln
@@ -1,11 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26730.15
+VisualStudioVersion = 15.0.27130.2010
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Maps", "Maps\Maps\Maps.csproj", "{4E3829FB-1D00-4FA7-A90B-F81FD4BB1034}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Maps.Tizen.Mobile", "Maps\Maps.Tizen.Mobile\Maps.Tizen.Mobile.csproj", "{475FD709-466A-4942-A36E-BB42504D41A9}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Maps.Tizen.Mobile", "Maps.Tizen.Mobile\Maps.Tizen.Mobile.csproj", "{649599CB-D235-40D3-AAD3-4C2A66CA2518}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,19 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4E3829FB-1D00-4FA7-A90B-F81FD4BB1034}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4E3829FB-1D00-4FA7-A90B-F81FD4BB1034}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4E3829FB-1D00-4FA7-A90B-F81FD4BB1034}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4E3829FB-1D00-4FA7-A90B-F81FD4BB1034}.Release|Any CPU.Build.0 = Release|Any CPU
-		{475FD709-466A-4942-A36E-BB42504D41A9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{475FD709-466A-4942-A36E-BB42504D41A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{475FD709-466A-4942-A36E-BB42504D41A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{475FD709-466A-4942-A36E-BB42504D41A9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{649599CB-D235-40D3-AAD3-4C2A66CA2518}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{649599CB-D235-40D3-AAD3-4C2A66CA2518}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{649599CB-D235-40D3-AAD3-4C2A66CA2518}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{649599CB-D235-40D3-AAD3-4C2A66CA2518}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {3B9E6463-D2C0-4D64-B8B1-3C588ADD6E92}
+		SolutionGuid = {A72715CE-A5B5-40EA-882B-B0345B4E41C8}
 	EndGlobalSection
 EndGlobal
