--- Mobile/Maps/Maps/Maps/Maps.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <!-- Include Nuget Package for Xamarin building -->
-  <ItemGroup>
-    <PackageReference Include="Tizen.NET" Version="5.0.0-preview1-00367" />
-    <PackageReference Include="Xamarin.Forms" Version="2.5.0.19271-pre2" />
-  </ItemGroup>
-
-</Project>
