+++ TV/Maps/Maps.Tizen.TV/shared/res/Maps.png
