+++ TV/Maps/Maps.Tizen.TV/tizen-manifest.xml
