+++ TV/Maps/Maps.sln
@@ -1,11 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26730.15
+VisualStudioVersion = 15.0.27130.2010
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Maps", "Maps\Maps\Maps.csproj", "{109D9BA4-6FC6-442A-94E5-45294822026F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Maps.Tizen.TV", "Maps\Maps.Tizen.TV\Maps.Tizen.TV.csproj", "{4F9036E8-EB26-47EA-921B-97EE3F8C37CC}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Maps.Tizen.TV", "Maps.Tizen.TV\Maps.Tizen.TV.csproj", "{DBA103FF-39B0-4E30-ACE9-C92879FE3784}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,19 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{109D9BA4-6FC6-442A-94E5-45294822026F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{109D9BA4-6FC6-442A-94E5-45294822026F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{109D9BA4-6FC6-442A-94E5-45294822026F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{109D9BA4-6FC6-442A-94E5-45294822026F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4F9036E8-EB26-47EA-921B-97EE3F8C37CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4F9036E8-EB26-47EA-921B-97EE3F8C37CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4F9036E8-EB26-47EA-921B-97EE3F8C37CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4F9036E8-EB26-47EA-921B-97EE3F8C37CC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DBA103FF-39B0-4E30-ACE9-C92879FE3784}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DBA103FF-39B0-4E30-ACE9-C92879FE3784}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DBA103FF-39B0-4E30-ACE9-C92879FE3784}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DBA103FF-39B0-4E30-ACE9-C92879FE3784}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A965AFF3-3D5A-4DB7-9FD9-E1EF3C2C2902}
+		SolutionGuid = {5B2650A4-459A-4E62-AB8B-DB42FC27F19D}
 	EndGlobalSection
 EndGlobal
