--- TV/Maps/Maps/Maps.Tizen.TV/Maps.Tizen.TV.cs
@@ -1,44 +1,0 @@
-/*
- * Copyright (c) 2016 Samsung Electronics Co., Ltd
- *
- * Licensed under the Flora License, Version 1.1 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-namespace Maps.Tizen.TV
-{
-    /// <summary>
-    /// Maps application main entry point.
-    /// </summary>
-    class Program : global::Xamarin.Forms.Platform.Tizen.FormsApplication
-    {
-        /// <summary>
-        /// Handle when your app creates.
-        /// </summary>
-        protected override void OnCreate()
-        {
-            base.OnCreate();
-            LoadApplication(new App());
-        }
-
-        /// <summary>
-        /// The entry point for the application.
-        /// </summary>
-        /// <param name="args"> A list of command line arguments.</param>
-        static void Main(string[] args)
-        {
-            var app = new Program();
-            global::Xamarin.Forms.Platform.Tizen.Forms.Init(app);
-            app.Run(args);
-        }
-    }
-}
