--- TV/Maps/Maps/Maps/Maps.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <!-- Include Nuget Package for Xamarin building -->
-  <ItemGroup>
-    <PackageReference Include="Tizen.NET" Version="4.0.0-preview1-00211" />
-    <PackageReference Include="Xamarin.Forms" Version="2.4.0.266-pre1" />
-  </ItemGroup>
-
-</Project>
